Addressing the popup window leak
Moved dismiss() call from onStop() to onPause(). Though this may
prevent the leak by guaranteeing that the popup logo will be dismissed
when the activity is hidden, the logo briefly fades out just before the
activity window is closed, which is a weird user experience. It may be best
to just use a regular app icon at this point.