Don't reload fragment when we don't have to
This still doesn't quite work yet. Some channels cause the drawer to
close when they shouldn't. Should find all usages of switchFragments and
figure out what they do exactly. Also still has that animation bug. Be
sure to overload onCreateAnimation in order to specify when to play the
animation or not.