Remove functionality to restore old fragments
This was never really something Android was meant to do and it was only
adding more instability. Now we just open the drawer on start. I've also
removed all usages of "val" and set it to error in the lombok config.
val is an interesting idea, but the implementation proves too finicky to
be consistently using.