Fix crash in places channel
Fix T68. Even though we close this issue this only raises more
questions. We may have fixed the crash here, but it cost us custom
animations in our fragments. Since this isn't the only bug in the
support library we've come across, we may want to look into some
solution or workaround for using them.