Fix Picasso image cropping
Fixes T71. I had to include a new library, picasso-transformations, in
order to get it to work. The library is really simple so if we don't
want to add another dependency, it would be pretty straightforward to
just rewrite the small part that we use. I also removed highlighting of
the drawer.