Add bookmarking
Users can now save any screen in the app that can be shared. This can be
done by selecting the bookmark option from the floating action button.
Changes under the hood include changes to the drawer adapter to allow
for bookmarks to be shown in the drawer and the bookmark page. Each
fragment controls its own toolbar because the toolbar can be changed for
tabs, search, etc.