Linking Synchronously
Summary:
Illustrate problems with linking

Sometimes we need to go to the internet to link. Sometimes we don't.
I've separated out the channels into those that need internet and those
that don't. This is not meant as a long term solution. We need to figure
out what to do with DTables, readers, and SoC to get them to not need
internet to link.

Test Plan: Ran it on my phone

Reviewers: jamchamb

Subscribers: mattro

Maniphest Tasks: T149

Differential Revision: http://phab.rutgers.edu/D21