Fixed a few unrelated bugs
There was a problem linking to DTables where it didn't give all
arguments needed. Picasso has problems with gifs for some reason, so for
now I check for gifs and don't load them. Hopefully this can be changed.
Also changed the linking scheme slightly. DTables no longer do special
sting escaping/formatting for making links. It was too confusing when
working with iOS, so we want to stay as close to the original titles as
possible.