Removed lombok
Lombok is nice and helped a lot to reduce boiler plate. Despite that we
removed lombok here, many more lines had to be added to compensate for
what it did. However, Android's new JACK system is incompatible with
lombok and I don't know if it will be any time soon. Also, you need a
intellij plugin to get it to work. Overall, I'm too afraid of being in
debt to lombok, so I decided to remove it now before it gets worse.