Start rewrite of bus, enable jack
Bus needs several things to be updated. First we need the bus API to use
retrofit throught. This seems mostly completed right now, but it needs
testing. Next all the bus fragments must be rewritten to use the new
api. This is mostly done, but BusAll still needs to be done. These also
require testing. Finally we need to be using RecyclerView throughout.
BusDisplay is the only part not using this.

Finally, jack has been enabled. I had problems getting it to work before
because it would run out of memory when dexing. I've configured it to
use it's own java process with a heap size of 4G. That seems to make it
work for now. This allows us native Java 8 features and a *much* smaller
apk size. I think it's down from ~45M to ~15M expanded. Big improvement!