Finish first draft bus rewrite
RecyclerView and Retrofit are used everywhere in bus. It still needs
some general cleanup since some methods are kinda garbo, but hopefully
we'll be rewriting bus soon anyway so it won't matter too much.

Also, surprise! Jack is being stupid again and was removed. I'd like to
add it back in eventually as it really improved APK size.