Refactor API
The API is now it's own module that has no knowledge of the app. There
are a few odds and ends that still need to be added like news stuff, but
for the most part everything is there.

There are two ways to set up the api. First is with simpleSetup where
you provide an OkHttp client and the API base url. The other is to just
provide a service that you created yourself (typically not what you want
to do). After that you can make static calls to the API.

I may change it to use a single instance class that just delegates to
the current objects, but we'll see what happens.