Add functionality for making alarms for when buses arrive
The user can press an icon in the BusDisplay which launches a Dialog
where they can configure how much time before the bus arrives at the
stop that was selected that they will be alerted.

This starts a service that sets an alarm that polls the Nextbus API on
an interval. When the lowest prediction time is below the user-set
threshold, the alarm is stopped and a high-priority notification is
shown to the user along with vibration and the selected alarm sound.

The notification can be refreshed or cancelled manually with buttons on
it.